Fixes route shape bug
Bug: A line was being drawn straight across the map from start of the route to the end.

This was caused by two of the same route shape getting concatenated onto a single arrival trip. Multiple trips can share the same trip shape. Previously, we were only looking for a change in the shape ID to determine when a particular shape ended, causing our bug when the next trip had the same shape id.  By also checking the trip ID we split trip shape and assign it to the next trip ID.